+++ src/week01/task3/FINRA__Mahmut.java
@@ -1,0 +1,32 @@
+package week01.task3;
+
+public class FINRA__Mahmut {
+    public static void main(String[] args) {
+
+        finra(20);
+    }
+    public static void finra(int num) {
+        String result = "";
+        for (int i = 1; i < num; i++) {
+            if (i % 5 == 0 && i % 3 == 0) {
+                result = "FINRA ";
+            } else if (i % 3 == 0) {
+                result = "FIN ";
+            } else if (i % 5 == 0) {
+                result = "RA ";
+            } else {
+                result = i + " ";
+            }
+            System.out.println(result);
+        }
+
+    }
+}
+
+/*
+3️⃣  Numbers -- FINRA
+Write a method that prints out the numbers from 1 to 30 but;
+for numbers which are multiple of 3 print "FIN" instead of the number,
+for numbers which are multiple of 5, print "RA" instead of the number,
+and for numbers which are multiple of both 3 and 5, print "FINRA" instead of the number.
+ */
